requirements.txt
Flask==3.0.3

Flask-SQLAlchemy==3.0.3
Werkzeug==3.0.3

bcrypt==4.0.1
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
gunicorn==21.2.0
reportlab==4.1
pandas==2.2.2
requests==2.31.0
yfinance==0.2.48
